server/requirements.txt
bs4
Flask
python-dotenv
